+++ RELEASES.rst
@@ -3557,8 +3557,9 @@ Planned
 * Include "end of input" in error message if SyntaxError occurs at end of
   file (GH-2152, GH-2165)
 
-* Fix behavior of proxied Array objects for Array.isArray(), duk_is_array(),
-  Object.prototype.toString(), and JSON.stringify() (GH-2041, GH-2175)
+* Fix behavior of proxied Array objects for: Array.isArray(), duk_is_array(),
+  Object.prototype.toString(), JSON.stringify(), Array.prototype.concat()
+  (GH-2041, GH-2175, GH-2176)
 
 * Fix a harmless assert in Math.atan2() when compiling with gcc -m32 (GH-2164)